% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.decrypt.R
\name{dbgap.decrypt}
\alias{dbgap.decrypt}
\title{Decrypt ncbi_enc files}
\usage{
dbgap.decrypt(file, key)
}
\arguments{
\item{file}{Pathway to the encrypted file or to the folder where the encrypted files are located}

\item{key}{Pathway to the key (*.ngc)}
}
\value{
Decrypt the file(s) and replace it in the same folder
}
\description{
This function decrypts dbGap files (*ncbi_enc) using your personnal key. Be careful, it can replace the settings of your "vdb-info" file
}
\author{
Gregoire Versmee, Laura Versmee
}
