% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbgap.expand.R
\name{dbgap.expand}
\alias{dbgap.expand}
\title{De-aggregate the variable in the datatables and create the first mapping file}
\usage{
dbgap.expand(phs, files, destination, study_name = phs)
}
\arguments{
\item{phs}{dbGap study ID in the for of phsxxxxxx}

\item{files}{Vector containing the pathways of the folders where the .txt files (or .txt.gz) are located for each consent groups that you want to extract}

\item{destination}{Pathway to the folder where you want to create the mapping file}

\item{study_name}{Nickname used for this study. Optional, otherwise will use the phs name}
}
\value{
a mapping file (0_map.csv), and a folder (study_tree) with the tree structure of your sandboxes. Also create a hidden file to copy oldmaps (.oldmaps)
}
\description{
This function extracts informations from .txt.gz files ind dbgap. It will de-agregate the datatables to create one csv file per variable with 2 columns (dbgap_ID and variable_name), and sort them in one folder per datatable. It also creates the first mapping file ("0_map.csv).
}
\author{
Gregoire Versmee, Laura Versmee
}
