\name{formula.character}
\alias{formula.character}
\title{Masked formula.character}
\usage{
  formula.character(x, env = parent.frame(), ...)
}
\arguments{
  \item{x}{see \code{formula.character}}

  \item{env}{see \code{formula.character}}

  \item{...}{see \code{formula.character}}
}
\description{
  Masked formula.character
}

