% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.oldmaps.R
\name{list.oldmaps}
\alias{list.oldmaps}
\title{Get the list of the saved old maps}
\usage{
list.oldmaps(mappath)
}
\arguments{
\item{mappath}{Pathway to the folder containing the mapping file}
}
\value{
Return a vector listing all the csv files containing saved maps
}
\description{
This function gets the list of the old maps saved in the ".oldmaps" folder.
}
\author{
Gregoire Versmee, Laura Versmee
}
