% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.oldmap.R
\name{look.oldmap}
\alias{look.oldmap}
\title{View an old mapping file on R}
\usage{
look.oldmap(mappath, old)
}
\arguments{
\item{mappath}{Pathway to the folder where your mapping file and your tree are located}

\item{oldcsv}{Name of the csv file that you want to use as a map. Old csv files can be found with the list.oldmaps function}
}
\value{
Return a view of your old mapping file as a data.frame
}
\description{
View an old mapping file on R
}
\author{
Gregoire Versmee, Laura Versmee
}
