% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapToTree.R
\name{MapToTree}
\alias{MapToTree}
\title{Change the structure of your tree folder using your mapping file}
\usage{
MapToTree(mappath)
}
\arguments{
\item{mappath}{Pathway to the folder where your mapping file and your tree are located}
}
\value{
Move and rename the files in the tree folder according to your mapping file
}
\description{
Change the structure of your tree folder using your mapping file
}
\author{
Gregoire Versmee, Laura Versmee
}
