% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.pop.R
\name{n.pop}
\alias{n.pop}
\title{Gets the population number of the study}
\usage{
n.pop(phs, consentgroups = TRUE)
}
\arguments{
\item{phs}{dbGap study ID (phs00xxxx, or 00xxxx, or xxx)}

\item{consentgroup}{if false, will return only the total number of participants}
}
\value{
a data.frame with 4 cols : name of the consent group, n male, n female, n total
}
\description{
This function extracts informations from data.dict.xml files from the dbgap ftp server to get the population characteristics. Works only for parents studies.
}
\author{
Gregoire Versmee, Laura Versmee
}
