% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover.map.R
\name{recover.map}
\alias{recover.map}
\title{Change the structure of your tree folder using an old mapping file}
\usage{
recover.map(oldcsv, mappath)
}
\arguments{
\item{oldcsv}{Name of the csv file that you want to use as a map. Old csv files can be found with the list.oldmaps function}

\item{mappath}{Pathway to the folder where your mapping file and your tree are located}
}
\value{
A new tree structure, according to the old map that you selected. The 0_map.csv is also changed for the one that you chose, and the old map is saved in the .oldmap folder
}
\description{
Change the structure of your tree folder using an old mapping file
}
\author{
Gregoire Versmee, Laura Versmee
}
