% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.expand.R
\name{table.expand}
\alias{table.expand}
\title{De-aggregate the variable in the datatables and create the first mapping file}
\usage{
table.expand(study_name, files, destination = getwd())
}
\arguments{
\item{study_name}{Nickname used for this study}

\item{files}{Vector containing the pathways to the table you want to expand}

\item{destination}{Pathway to the folder where you want to create the mapping file}
}
\value{
a mapping file (0_map.csv), and a folder (study_tree) with the tree structure of your sandboxes. Also create a hidden file to copy oldmaps (.oldmaps)
}
\description{
This function extracts informations from .txt.gz files ind dbgap. It will de-agregate the datatables to create one csv file per variable with 2 columns (dbgap_ID and variable_name), and sort them in one folder per datatable. It also creates the first mapping file ("0_map.csv).
}
\author{
Gregoire Versmee, Laura Versmee
}
