% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeToMap.R
\name{TreeToMap}
\alias{TreeToMap}
\title{Change your mapping file according to the tree structure}
\usage{
TreeToMap(mappath)
}
\arguments{
\item{mappath}{Pathway to the folder where your mapping file and your tree are located}
}
\value{
A new mapping file "0_map.csv" according to your tree structure. The old map is moved to a hidden folder ".oldmaps"
}
\description{
Change your mapping file according to the tree structure
}
\author{
Gregoire Versmee, Laura Versmee
}
